requirements.txt
APScheduler~=3.11.1
SQLAlchemy~=2.0.45
fastapi~=0.125.0
aiohttp~=3.13.2
beautifulsoup4~=4.14.3
pydantic~=2.12.5
alembic~=1.17.2
python-decouple